+++ front/src/modal/layout/CreateEventModalLayout/index.jsx
@@ -4,20 +4,28 @@ import CreateEventModal from '../../component/CreateEventModal';
 import EventColorModal from '../../component/EventColorModal';
 import ListModal from '../../component/ListModal';
 import CustomAlertModal from '../../component/CustomAlertModal';
+import MiniCalendarModal from '../../component/MiniCalendarModal';
+import TimeListModal from '../../component/TimeListModal';
 
 import {
   CreateEventModalContext,
   EventColorModalContext,
   EventInfoListModalContext,
   EventCustomAlertModalContext,
+  EventDateModalContext,
 } from '../../../context/EventModalContext';
 import useEventModal from '../../../hooks/useEventModal';
+import { useDispatch } from 'react-redux';
+import { updateNewEventTime } from '../../../store/newEvent';
 
 const Index = ({ children }) => {
   const createEventModal = useEventModal();
   const eventColorModal = useEventModal();
   const eventInfoModal = useEventModal();
   const eventCustomAlertModal = useEventModal();
+  const miniCalendarModal = useEventModal();
+  const startTimeListModal = useEventModal();
+  const endTimeListModal = useEventModal();
 
   function hideCreateEventModal() {
     createEventModal.hideModal();
@@ -57,6 +65,28 @@ const Index = ({ children }) => {
     ...eventInfoModal,
     showModal: showEventInfoModal,
   };
+
+  const eventDateModalContextData = {
+    miniCalendarModal,
+    startTimeListModal,
+    endTimeListModal,
+  };
+
+  const dispatch = useDispatch();
+  function onClickTimeItem(e, type) {
+    const date = new Date(+e.target.dataset.value);
+    dispatch(
+      updateNewEventTime({
+        type: 'startTime',
+        minute: date.getMinutes(),
+        hour: date.getHours(),
+      }),
+    );
+    if (type === 'startTime') startTimeListModal.hideModal();
+    if (type === 'endTime') endTimeListModal.hideModal();
+    e.stopPropagation();
+  }
+
   return (
     <CreateEventModalContext.Provider value={createEventModalContext}>
       {createEventModal.isModalShown && (
@@ -65,21 +95,48 @@ const Index = ({ children }) => {
             <EventCustomAlertModalContext.Provider
               value={eventCustomAlertModal}
             >
-              <CreateEventModal>
-                {eventColorModal.isModalShown && <EventColorModal />}
-                {eventInfoModal.isModalShown && (
-                  <ListModal
-                    hideModal={eventInfoModal.hideModal}
-                    modalData={eventInfoModal.modalData}
-                  />
-                )}
-                {eventCustomAlertModal.isModalShown && (
-                  <CustomAlertModal
-                    hideModal={eventCustomAlertModal.hideModal}
-                    modalData={eventCustomAlertModal.modalData}
-                  />
-                )}
-              </CreateEventModal>
+              <EventDateModalContext.Provider value={eventDateModalContextData}>
+                <CreateEventModal>
+                  {eventInfoModal.isModalShown && (
+                    <ListModal
+                      hideModal={eventInfoModal.hideModal}
+                      modalData={eventInfoModal.modalData}
+                    />
+                  )}
+
+                  {eventColorModal.isModalShown && <EventColorModal />}
+
+                  {eventCustomAlertModal.isModalShown && (
+                    <CustomAlertModal
+                      hideModal={eventCustomAlertModal.hideModal}
+                      modalData={eventCustomAlertModal.modalData}
+                    />
+                  )}
+
+                  {miniCalendarModal.isModalShown && (
+                    <MiniCalendarModal
+                      hideModal={miniCalendarModal.hideModal}
+                      modalData={miniCalendarModal.modalData}
+                    />
+                  )}
+
+                  {startTimeListModal.isModalShown && (
+                    <TimeListModal
+                      hideModal={startTimeListModal.hideModal}
+                      modalData={startTimeListModal.modalData}
+                      onClickItem={e => onClickTimeItem(e, 'startTime')}
+                    />
+                  )}
+
+                  {endTimeListModal.isModalShown && (
+                    <TimeListModal
+                      hideModal={endTimeListModal.hideModal}
+                      modalData={endTimeListModal.modalData}
+                      onClickItem={e => onClickTimeItem(e, 'endTime')}
+                    />
+                  )}
+                </CreateEventModal>
+              </EventDateModalContext.Provider>
             </EventCustomAlertModalContext.Provider>
           </EventInfoListModalContext.Provider>
         </EventColorModalContext.Provider>
